Improve our handling of page dir values per thread.
First, fixup the loading of swapper_pg_dir -- newer kernels don't have
it.  If it's not there, read init_mm.pgd instead.

Then, when loading threads from memory, if the thread is a kernel thread,
instead of leaving the thread's pgd value empty, set it to whatever we
got from the above detection.

Finally, also save each thread's current pgd value in our personality
thread info, and invalidate it at the right time to force it to be
reloaded.  No sense in re-translating it (v2p) each time we need it,
even if we're caching the translation in a mem backend...