Make sure threads sleeping in schedule() on x86_64 have an IP.
Linux saves %eip on x86 systems in context switch in thread_struct,
which is triggered when a thread calls schedule().

Linux on x86_64 doesn't have to save %rip in context switch after
schedule(); the context switch is effected by swapping new/old %rsp
(and of course all the other state too).  This means there is no
saved %rip for a thread that has "stopped" executing during context
switch.  This meant we could not backtrace sleeping threads.  Well,
there is one exception to this rule -- threads that were preempted
in the kernel will of course have %rip on the stack.  We already
handle that one.

So these sleeping threads that are waiting to be context-switched in
might not have saved %rip, but we know where it is.  Basically, for
preexisting threads, we choose to set the %rip at the point in context
switch where the scheduler has just saved its %rsp into thread_struct;
so the next instruction (when the sleeping thread is resumed) will be
to set the current %rsp to the value that was saved in thread_struct.
This seems to be the correct point because we want our backtracer to
see a "correct" stack (plus, the rest of schedule() and __switch_to
happen on the stack of the incoming thread).

But the situation is different for newly-forked threads.  We detect
those the same way the kernel does -- _TIF_FORK set in
thread_info->flags.  In this case, the thread immediately returns via
ret_from_fork as it starts its execution.  So that is what we set the
IP to in this case.

A backtrace in every pot.