Fix active thread_entry probing when copy_process is not probe-able.
On some Linuxes, copy_process (the best previous point we found to
hook to catch the new process after it's been fully created, or nearly
so) can't be probed the way we need.  Basically, it can get inlined,
and we don't have function entry/exit probes for inlined functions
(and we never will, because the inlined functions of course don't
have well-defined exit points anymore...).

So, we ifdef out the copy_process stuff, and instead place probes
on wake_up_new_task instead.  Not sure why I didn't use this initially,
but it seems to be a good choice.  wake_up_new_task basically puts the
task onto the run queues, so it's a great choice for us.