Big speedup of the ELF range handling code.
For quite some time now, the ELF symbol loader has attempted to
assign sizes to 0-byte length symbols.  We do this to create an
indexed (by base addr) list of symbol *ranges*, which we can then
use to search the program's address space.  We can debate the
merits of this, but anyway...

Before, I had written some complex code that would call the very
expensive clrange_update_end to change the range of an already-
inserted symbol.  This is quite complicated, and it was incredibly
slow.

So what this new code does is save 0-byte length symbols until a
final pass at the end, and it does not add them to the range data
structure as it is reading the ELF symbol tables, like before.

Thus there is no update to be done; we can just add the symbols to
the range structure if we can infer a length based on the following
symbol or the end of the ELF section.  Note we do this in reverse
order, of course, because the "following" symbol for any given
0-byte length symbol X might not have been added yet!  So we have to
go backwards.

This is much faster, O(n) instead of O(n**2), if I recall correctly
what it was.  Silly me.  Saves *seconds* on massive kernel debugfiles.