Add some real examples, and an example Makefile, for users.
This Makefile uses pkg-config based on the .pc files installed for
the target and vmilib sub-libraries, and assumes those files were
installed in @prefix@ -- which is how things should be.

Users should setup their Makefiles like this to build and link
against the installed libraries.

Also, I wrote two quick examples, demonstrating the use of the
ptrace driver specifically, and the target library generally.

ptrace.c allows you to attach to a single process and probe it.

ptrace-multi-process.c allows you to attach to multiple processes,
and probe them selectively.  This shows how to use target_poll instead
of target_monitor(); etc.

ptrace.c is highly-commented; ptrace-multi-process.c assume you've
read the comments in ptrace.c ; but it does add a few comments
explaining some of the differences.