Never cache writable-mapped pages.
We were caching mmaps of both read-only and writable pages (so we could
read and/or write to pages).  This turned out to cause some problems for
the hypervisor.  After mulling over the code, I believe this happens for
PV domains for several reasons.

First, PV guests are aware of the MFNs they have been allocated, and when
they construct their page tables, they construct the entries in terms of
MFNs.  Then, when they load cr3, the hypervisor validates them.  What this
means is that the OS believes it is the sole owner of MFNs, and can thus
change their protection whenever it wants (or use them as page table
pages -- and Xen insists that page tables be marked read-only, of course).

Given this situation... if we mmap a physical page writable because at the
time it really is writable and we need to write to it, and then hold onto
the mmap -- if the PV guest later decides to reuse that page for a read-only
purpose (i.e., as a page table page, or anything else), there will be a
conflict in the hypervisor when the page type is changed, because the
hypervisor will see that the page has a nonzero refcnt, and page types can
only be changed when the refcnt is zero.

At least that's what I think I understand or surmise.

VMI fundamentally cannot deal with this unless we snoop on OS page tables
too, to see when PV MFNs are changed to a different type -- and of course
we also can't see when they change protection either (well, we could, but
the resulting probe overhead is far too high -- and probably even shadow
page table tricks are too high too -- but we could at least try that).
Anyway, for now, we don't bother doing any of this stuff, because writes
are relatively rare compared to reads.  It is tempting to wish we could
allow write caching of pages between subsequent interrupts (i.e., so we
could flip software breakpoints cheapest) -- but this will certainly fail
eventually since sometimes breakpoints can be hit in one OS thread just
prior to a context switch -- and in that case we reexecute the breakpoint
next time the previous thread is reentered.  So even limited caching is
nontrivial and prone to racy behavior.

Anyway, this approach works plenty fine.