Lots of little CFA tweaks; one bigger tweak for inline instances.
Mostly, this helps us handle the cases where there is no symbol
associated with the current lctxt; or when it has no name; or even
if the lops provide no getsymbol() for us.

More important, do not just start our lookup in the parent function
if the symbol is an inline instance!  Sometimes an inline instance
could have FDE info associated with it directly; and in any case,
there is already code recursively checking up the parent hierarchy
for a scope range that does have FDE info.  This gets rid of some
misleading warnings too.