Fix singlestep handling for xen/os-process drivers.
For the Xens that don't set debug registers correctly on singlestep
trap (and when combined with an overlay process target), we used to
have the overlay target detect that by looking into the base's data
structures.  With the refactor, that went away.  However, somehow
the refactor, in one of its code paths for notifying the overlay,
didn't carry this change through.

So, now, when the underlying Xen target doesn't know what kind of
debug exception it is based on debug status registers, it sets the
exception flags to EXCEPTION_BREAKPOINT, and maybe adds
EXCEPTION_SINGLESTEP if the debug status reg was set; and maybe adds
EXCEPTION_SINGLESTEP_CMD if we can tell that a single step had been
commanded for the thread.

Then, I rearranged the os-process driver's handling to check for
breakpoint hit first, to handle the case where we might be stepping
already and then hit a breakpoint.  (The default probe bp_handler
checks for this case and notifies the step handler before it proceeds
to handle the breakpoint.)  Then it also now properly handles the
inferred single step case.

This also gets rid of the (stupid) EXCEPTION_BREAKPOINT_STEP flag,
and adds the much more useful EXCEPTION_SINGLESTEP_CMD flag to
highlight the case where we had commanded a step.