Add support to attach to multiple Xen VMs in the same Stackdb program.
All I had to do was move client-side connections to the multiplexer into
the per-target driver state; each VM gets a connection to the multiplexer.
Don't have to worry about scalability of that...

If you specify --no-use-multiplexer, you are still limited to attaching
to one VM per Xen phys host.