QEMU shim: cleanup dangling hugetlbfs files we skipped unlink of.
NB: this only works when QEMU shuts itself down, via SIGINT, SIGHUP,
or SIGTERM.  Happily, the signal-handling mechansim is also used if
the machine shuts itself down inside the emulation (i.e., shutdown -h).

Obviously it won't work if the emulator crashes or is SIGKILL'd; in
that case, you still have to cleanup to avoid exhaustion of your
hugetlbfs.

(In the future, we could intercept the call to open(2) that opens
the hugetlbfs memory-backing file and add the O_TMPFILE flag (to
ensure that the file is removed as soon as no one has it open),
but that only works since kernel 3.11 .)