Parallelize the build where we can.
On my quad-core desktop, it's good for a 2.5x speedup, from
over 2 minutes to just over 45 seconds.

Basically, we can only parallelize inside the subdirs because
they have tight deps on each other (we would have to decouple
make all into make obj and make link, etc, which of course we
won't do...).  And then the java stuff is complicated in the
same way, and even worse because of how the client stubs and
skeletons are built in their hacky way.