In some key places, use a dynamic buf with value_snprintf.
Sometimes we have to get the whole value string, or not at all
(i.e., in output we send to the xml server, which expects
well-formed strings).  So, if value_snprintf couldn't fit its
output in our static buf, use its estimated necessary size to
dynamically allocate a buf to use instead.