Fix up the ASM stuff to build against an installed CLIPS.
Now it's --disable-asm; asm will be built by default if we
can find an installed clips; and we support --with-clips to
help find that installed lib if necessary.  --with-clipssrc is
still available, and the right include paths and headers are used.

Oh, and fix up some xen turds in the build, and make the policy_engine
available for KVM, not only Xen.