When getting page prot bits in Linux targets, get arch-indep version.
When loading vm_area_structs, we used to return the arch-specific
vm_page_prot field.  That field is arch-specific, meaning it is ready to
be loaded into a PTE.  The non-arch-specific field is vm_flags.  So, we
really want to return the non-arch-specific since we're at the semantic
layer.  If this causes problems, we can add an arch-specific field, too.